moves images folder into assets folder and refactors image paths
