+++ src/index.js
@@ -7,20 +7,6 @@ import 'adaptiv-ui/css/main.css'
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-// Auth0 imports
-// import { Auth0Provider } from "./components/auth/react-auth0-spa";
-// import config from "./components/auth/auth_config.json";
-import history from "./utils/History";
-
-// Routes the user to dashboard upon login
-const onRedirectCallback = appState => {
-  history.push(
-    appState && appState.targetUrl
-      ? appState.targetUrl
-      : window.location.pathname
-  );
-};
-
 
 ReactDOM.render(
   <AppWrapper bg='white' >
