+++ src/pages/DashRouter.js
@@ -1,11 +1,13 @@
-import React from "react";
-import { useAuth0 } from "../components/auth/react-auth0-spa";
-import config from "../components/auth/auth_config.json";
-import AdminDashboard from "./admin/AdminDashboard";
-import UserDashboard from "./users/UserDashboard";
-import { Flex, Box } from "adaptiv-ui";
+// React imports
+import React from 'react';
 
-// example Profile page
+// Auth0 imports
+import { useAuth0 } from '../components/auth/react-auth0-spa';
+import config from '../components/auth/auth_config.json';
+
+// Component imports
+import AdminDashboard from './admin/AdminDashboard';
+import UserDashboard from './users/UserDashboard';
 
 const DashRouter = () => {
   const { loading, user } = useAuth0();
@@ -16,7 +18,7 @@ const DashRouter = () => {
     return <div>Loading...</div>;
   }
 
-  return user[config.roleUrl].includes("Admin") ? (
+  return user[config.roleUrl].includes('Admin') ? (
     <div>
       <AdminDashboard user={user} />
     </div>
